chore(packages): Update package definitions for moved classes
