+++ src/main/java/edu/ntnu/idi/idatt/factory/PlayerFactory.java
@@ -1,4 +1,4 @@
-package edu.ntnu.idi.idatt.model.factory;
+package edu.ntnu.idi.idatt.factory;
 
 import edu.ntnu.idi.idatt.model.Player;
 import edu.ntnu.idi.idatt.model.PlayerTokenType;
